Cookies configured, now the quantity and productId are passed as Javascript Objects (see console logs for examples)

+++ store/models.py
@@ -15,7 +15,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField(null=True)
+    price = models.DecimalField(max_digits=7, null=True, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
